PhysicalPoint & LogicalPoint (#349)
* LogicalPoint

* Fix flipped y

* Update changes in three-d-asset

* Fix mandelbrot example

* Use PhysicalPoint in 2d geometry

* Docs

* More conversions

* Use three-d-asset version from github

* Fix clippy warnings

* device_pixel_ratio is f32

* Use github version of three-d-asset in all examples

* Fix egui-gui + examples

* Fix tests

+++ examples/lighting/src/main.rs
@@ -216,7 +216,7 @@ pub async fn run() {
                     ui.radio_value(&mut material_type, MaterialType::Depth, "Depth");
                     ui.radio_value(&mut material_type, MaterialType::Orm, "ORM");
                 });
-                panel_width = gui_context.used_rect().width() as f64;
+                panel_width = gui_context.used_rect().width();
             },
         );
