+++ src/renderer/control/camera_control.rs
@@ -147,35 +147,35 @@ impl CameraControl {
         change
     }
 
-    fn handle_action(&mut self, camera: &mut Camera, control_type: CameraAction, x: f64) -> bool {
+    fn handle_action(&mut self, camera: &mut Camera, control_type: CameraAction, x: f32) -> bool {
         match control_type {
             CameraAction::Pitch { speed } => {
-                camera.pitch(radians(speed * x as f32));
+                camera.pitch(radians(speed * x));
             }
             CameraAction::OrbitUp { speed, target } => {
-                camera.rotate_around_with_fixed_up(&target, 0.0, speed * x as f32);
+                camera.rotate_around_with_fixed_up(&target, 0.0, speed * x);
             }
             CameraAction::Yaw { speed } => {
-                camera.yaw(radians(speed * x as f32));
+                camera.yaw(radians(speed * x));
             }
             CameraAction::OrbitLeft { speed, target } => {
-                camera.rotate_around_with_fixed_up(&target, speed * x as f32, 0.0);
+                camera.rotate_around_with_fixed_up(&target, speed * x, 0.0);
             }
             CameraAction::Roll { speed } => {
-                camera.roll(radians(speed * x as f32));
+                camera.roll(radians(speed * x));
             }
             CameraAction::Left { speed } => {
-                let change = -camera.right_direction() * x as f32 * speed;
+                let change = -camera.right_direction() * x * speed;
                 camera.translate(&change);
             }
             CameraAction::Up { speed } => {
                 let right = camera.right_direction();
                 let up = right.cross(camera.view_direction());
-                let change = up * x as f32 * speed;
+                let change = up * x * speed;
                 camera.translate(&change);
             }
             CameraAction::Forward { speed } => {
-                let change = camera.view_direction() * speed * x as f32;
+                let change = camera.view_direction() * speed * x;
                 camera.translate(&change);
             }
             CameraAction::Zoom {
@@ -184,7 +184,7 @@ impl CameraControl {
                 min,
                 max,
             } => {
-                camera.zoom_towards(&target, speed * x as f32, min, max);
+                camera.zoom_towards(&target, speed * x, min, max);
             }
             CameraAction::None => {}
         }
